Refactor for packge to get working
Had to move websocket for getting data to different process/
service as a new process is tarted for each connection to the server.
Thus, the socket runs as service to get data and save to file
while the web server reads the file every so often and plots
the data.

v0.1.0